requirements.txt
Django==1.7
psycopg2
dj-static==0.0.6
django-filter==0.7
django-tastypie==0.11.0
djangorestframework==2.4.3
drf-nested-routers
djangorestframework-bulk
gunicorn==18.0
python-dateutil==1.5
virtualenv==1.10.1
django-admin-bootstrapped
django-keyedcache